requirements.txt
playwright==1.42.0
llama-cpp-python==0.2.38
pyyaml==6.0.1
fastapi==0.109.0
uvicorn==0.27.0
python-owasp-zap-v2.4==0.0.20
pytest==7.4.3
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0